reuse wheel handlers and apply to comp workspace
