+++ src/project/composition/handlers/dragProjectCompWorkspaceToArea.tsx
@@ -1,13 +1,13 @@
-import { getActionState } from "~/state/stateUtils";
-import { requestAction } from "~/listener/requestAction";
-import { getDistance } from "~/util/math";
 import { areaActions } from "~/area/state/areaActions";
-import { AreaType } from "~/constants";
 import { areaInitialStates } from "~/area/state/areaInitialStates";
-import { getAreaRootViewport } from "~/area/util/getAreaViewport";
 import { computeAreaToViewport } from "~/area/util/areaToViewport";
 import { getAreaToOpenTargetId } from "~/area/util/areaUtils";
-import { CompositionWorkspaceAreaState } from "~/composition/workspace/compositionWorkspaceAreaReducer";
+import { getAreaRootViewport } from "~/area/util/getAreaViewport";
+import { CompositionWorkspaceAreaState } from "~/composition/workspace/compWorkspaceAreaReducer";
+import { AreaType } from "~/constants";
+import { requestAction } from "~/listener/requestAction";
+import { getActionState } from "~/state/stateUtils";
+import { getDistance } from "~/util/math";
 
 interface Options {
 	compositionId: string;
